requirements.txt
Django==5.0.7
djangorestframework==3.15.2
celery==5.3.4
redis==5.0.1
twilio==8.10.0
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
django-cors-headers==4.3.1
dj-database-url==2.1.0
whitenoise==6.6.0
